Moved reports-backend/src/main/java/org/jboss/da/reports/api/SCMLocator.java -> reports-model/src/main/java/org/jboss/da/reports/model/api/SCMLocator.java
