+++ bc-backend/src/main/java/org/jboss/da/bc/api/ProductBuildConfigurationGenerator.java
@@ -5,7 +5,7 @@ import org.jboss.da.bc.model.backend.ProductGeneratorEntity;
 import org.jboss.da.common.CommunicationException;
 import org.jboss.da.communication.pnc.api.PNCRequestException;
 import org.jboss.da.communication.pom.PomAnalysisException;
-import org.jboss.da.reports.api.SCMLocator;
+import org.jboss.da.reports.model.api.SCMLocator;
 
 import java.util.Optional;
