+++ bc-backend/src/main/java/org/jboss/da/bc/api/ProjectBuildConfigurationGenerator.java
@@ -6,7 +6,7 @@ import org.jboss.da.bc.model.backend.ProjectGeneratorEntity;
 import org.jboss.da.common.CommunicationException;
 import org.jboss.da.communication.pnc.api.PNCRequestException;
 import org.jboss.da.communication.pom.PomAnalysisException;
-import org.jboss.da.reports.api.SCMLocator;
+import org.jboss.da.reports.model.api.SCMLocator;
 
 /**
  *
