+++ bc-backend/src/main/java/org/jboss/da/bc/impl/BuildConfigurationGeneratorImpl.java
@@ -23,9 +23,9 @@ import org.jboss.da.bc.model.backend.ProjectHiearchy;
 import org.jboss.da.common.CommunicationException;
 import org.jboss.da.communication.pnc.api.PNCRequestException;
 import org.jboss.da.communication.pom.PomAnalysisException;
-import org.jboss.da.reports.api.SCMLocator;
 import org.jboss.da.reports.backend.api.DependencyTreeGenerator;
 import org.jboss.da.reports.backend.api.GAVToplevelDependencies;
+import org.jboss.da.reports.model.api.SCMLocator;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
