+++ bc-backend/src/main/java/org/jboss/da/bc/model/backend/GeneratorEntity.java
@@ -4,7 +4,7 @@ import lombok.Getter;
 import lombok.Setter;
 
 import org.jboss.da.model.rest.GAV;
-import org.jboss.da.reports.api.SCMLocator;
+import org.jboss.da.reports.model.api.SCMLocator;
 
 /**
  *
