+++ bc-backend/src/main/java/org/jboss/da/bc/model/backend/ProjectGeneratorEntity.java
@@ -3,7 +3,7 @@ package org.jboss.da.bc.model.backend;
 import lombok.ToString;
 
 import org.jboss.da.model.rest.GAV;
-import org.jboss.da.reports.api.SCMLocator;
+import org.jboss.da.reports.model.api.SCMLocator;
 
 /**
  *
