+++ bc-rest/src/main/java/org/jboss/da/bc/facade/AbstractBuildConfigurationsFacade.java
@@ -19,7 +19,7 @@ import org.jboss.da.common.CommunicationException;
 import org.jboss.da.communication.pnc.api.PNCRequestException;
 import org.jboss.da.communication.pom.PomAnalysisException;
 import org.jboss.da.model.rest.GAV;
-import org.jboss.da.reports.api.SCMLocator;
+import org.jboss.da.reports.model.api.SCMLocator;
 import org.slf4j.Logger;
 
 /**
