+++ bc-rest/src/main/java/org/jboss/da/bc/facade/BuildConfigurationsProjectFacade.java
@@ -9,7 +9,7 @@ import org.jboss.da.bc.model.rest.ProjectInfoEntity;
 import org.jboss.da.common.CommunicationException;
 import org.jboss.da.communication.pnc.api.PNCRequestException;
 import org.jboss.da.communication.pom.PomAnalysisException;
-import org.jboss.da.reports.api.SCMLocator;
+import org.jboss.da.reports.model.api.SCMLocator;
 import org.slf4j.Logger;
 
 import javax.enterprise.context.ApplicationScoped;
